pyproject.toml
[project]
name = "cg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyopengl>=3.1.9",
]
